Factor out required function for reuse
